requirements.txt
-i https://pypi.python.org/simple
a2wsgi==1.10.10; python_full_version >= '3.8.0'
alembic==1.16.4; python_version >= '3.9'
aniso8601==10.0.1
anyio==4.9.0; python_version >= '3.9'
app-common-python==0.2.8
apscheduler==3.11.0; python_version >= '3.8'
asgiref==3.9.1; python_version >= '3.9'
async-timeout==5.0.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
backoff==2.2.1; python_version >= '3.7' and python_version < '4.0'
backports-datetime-fromisoformat==2.0.3; python_version >= '3.1'
backports.entry-points-selectable==1.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
boto3==1.39.8; python_version >= '3.9'
botocore==1.39.8; python_version >= '3.9'
cachelib==0.13.0; python_version >= '3.8'
certifi==2025.7.14; python_version >= '3.7'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
confluent-kafka==2.10.1; python_version >= '3.7'
connexion[flask,swagger-ui,uvicorn]==3.2.0; python_version >= '3.8' and python_version < '4.0'
distlib==0.4.0
exceptiongroup==1.3.0; python_version >= '3.7'
fcache==0.6.0; python_version >= '3.9'
filelock==3.18.0; python_version >= '3.9'
flask[async]==3.1.1; python_version >= '3.9'
flask-caching==2.3.1; python_version >= '3.8'
flask-migrate==4.1.0; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
flask-unleash==2.1.0
gunicorn==23.0.0; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
honcho==2.0.0
httpcore==1.0.9; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.7.0; python_version >= '3.9'
inflection==0.5.1; python_version >= '3.5'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonify==0.5
jsonschema==4.24.0; python_version >= '3.9'
jsonschema-path==0.3.4; python_full_version >= '3.8.0' and python_full_version < '4.0.0'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
lazy-object-proxy==1.11.0; python_version >= '3.9'
logstash-formatter==0.5.17
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==4.0.0; python_version >= '3.9'
mmh3==5.1.0; python_version >= '3.9'
openapi-schema-validator==0.6.3; python_full_version >= '3.8.0' and python_full_version < '4.0.0'
openapi-spec-validator==0.7.2; python_full_version >= '3.8.0' and python_full_version < '4.0.0'
packaging==25.0; python_version >= '3.8'
pathable==0.4.4; python_full_version >= '3.7.0' and python_full_version < '4.0.0'
platformdirs==3.11.0; python_version >= '3.7'
prance==23.6.21.0; python_version >= '3.8'
prometheus-client==0.22.1; python_version >= '3.9'
prometheus-flask-exporter==0.23.2
psycopg2==2.9.10; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.1; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
ratelimit==2.2.1
redis==6.2.0; python_version >= '3.9'
referencing==0.36.2; python_version >= '3.9'
requests==2.32.4; python_version >= '3.8'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rpds-py==0.23.1; python_version >= '3.9'
ruamel.yaml==0.18.14; python_version >= '3.8'
ruamel.yaml.clib==0.2.12; python_version >= '3.9'
s3transfer==0.13.0; python_version >= '3.9'
segment-analytics-python==2.3.4; python_full_version >= '3.9.0'
semver==3.0.4; python_version >= '3.7'
setuptools==80.9.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.41; python_version >= '3.7'
starlette==0.47.1; python_version >= '3.9'
swagger-ui-bundle==1.1.0; python_version >= '3.7'
tomli==2.0.2; python_version >= '3.8'
typing-extensions==4.14.1; python_version >= '3.9'
tzlocal==5.3.1; python_version >= '3.9'
ujson==5.10.0; python_version >= '3.8'
unleashclient==5.12.2; python_version >= '3.7'
urllib3==1.26.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
validators==0.35.0; python_version >= '3.9'
virtualenv==20.32.0; python_version >= '3.8'
watchfiles==1.1.0; python_version >= '3.9'
watchtower==3.4.0; python_version >= '3.8'
websockets==15.0.1; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
zipp==3.23.0; python_version >= '3.9'

# OpenTelemetry instrumentation
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-kafka-python==0.48b0
opentelemetry-instrumentation-psycopg2==0.48b0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-jaeger==1.27.0
opentelemetry-propagator-b3==1.27.0
